Remove Table Sync Supressions
This commit remove the now working table supressions that were tested as part of #54. We still supress data_browser because that is a meta table and the means_by_traveltime_to_work_acs_bg because we do not know why that table is failing to sync. However all the other tables sync’d correctly with a manual POST sync request.