Update Data Title
Our dataset has annual data points for monthly data, not weekly. So this updates the title to reflect the fact.

EOWLD publishes monthly employment figures throughout the year and then averages them to an annual monthly average at the end of the year. so, in any given month there are x people employed on average throughout the year.

Co-authored-by: Henry Quinn <henryquinniv@gmail.com>